Fixed signed vs unsigned comparison warning generated when compiling with msys.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2011/selfshadow@36639 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/rendermanager/osm/shadow_osm.h
@@ -408,7 +408,7 @@ namespace CS
               typename RenderTree::ContextNode* shadowMapCtx = 
                 renderTree.CreateContext (newRenderView);
 
-              for (size_t t = 0; t < persist.mrt; t++)
+              for (int t = 0; t < persist.mrt; t++)
                 if (texHandles[t] != 0)
                 {
                   renderTree.AddDebugTexture (texHandles[t]);
